backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.19
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pytest==8.3.4
httpx==0.28.1
authlib==1.6.5
itsdangerous==2.2.0
email-validator==2.2.0
google-cloud-firestore==2.14.0
google-cloud-datastore==2.19.0
google-auth==2.27.0
